requirements.txt
psycopg2-binary>=2.9
sqlalchemy>=2.0
geoalchemy2>=0.15
ultralytics>=8.3
torch>=2.0
mercantile>=1.2
rasterio>=1.3
shapely>=2.0
pyproj>=3.6
streamlit>=1.37
streamlit-folium>=0.22
folium>=0.17
pydeck>=0.8
pillow>=10.0
numpy>=1.26
pandas>=2.0
geopandas>=1.0
tqdm>=4.66
python-dotenv>=1.0
